data-loader/requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
numpy==1.26.2

